+++ src/za/co/fynbos/strategy/impl/CreditCardStrategy.java
@@ -17,7 +17,7 @@ public class CreditCardStrategy implements PaymentStrategy{
 	}
 	
 	@Override
-	public void pay(int amount) {
+	public void pay(double amount) {
 		System.out.println(amount +" paid with credit/debit card");
 	}
